Add Unit Test For Report Builder Linked List
Adds unit test for report builder linked list to ensure that NAN values
are populated as expected as well as injected data.
