+++ test/src/apps/bridge/CMakeLists.txt
@@ -30,3 +30,31 @@ target_sources(sm_config_crc_list_tests
 target_link_libraries(sm_config_crc_list_tests gmock gtest_main)
 
 add_test(NAME sm_config_crc_list_tests COMMAND sm_config_crc_list_tests)
+
+add_executable(report_builder_tests)
+target_include_directories(report_builder_tests
+    PRIVATE
+    ${TEST_DIR}/header_overrides
+    ${SRC_DIR}/apps/bridge
+    ${SRC_DIR}/apps/bridge/sensor_drivers
+    ${SRC_DIR}/lib/bm_core/bcmp
+    ${SRC_DIR}/lib/bm_core/bm_common_messages
+    ${SRC_DIR}/lib/bm_core/common
+    ${SRC_DIR}/lib/bm_core/middleware
+    ${SRC_DIR}/lib/bm_core/test
+    ${SRC_DIR}/lib/common
+    ${SRC_DIR}/third_party/FreeRTOS/Source/include
+    ${TEST_DIR}/mocks
+    ${TEST_DIR}/third_party/fff
+    ${SRC_DIR}/third_party/tinycbor/src
+)
+
+target_sources(report_builder_tests
+    PRIVATE
+    ${TEST_DIR}/stubs/mock_bridgeLog.cpp
+    ${SRC_DIR}/apps/bridge/reportBuilderList.cpp
+    report_builder_ut.cpp)
+
+target_link_libraries(report_builder_tests gmock gtest_main)
+
+add_test(NAME report_builder_tests COMMAND report_builder_tests)
